added fetchUserToken function to helpers folder
